REFACTOR: NAV-104 - Move StopConnection logic to object itself instead of DtoMapper.
